backend/requirements.txt
flask
python-constraint
networkx
fastapi
pydantic